requirements.txt
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
gunicorn==20.1.0
mediapipe==0.8.3.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.1.2
psycopg2==2.7.7
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
